Fix: PR #96 removed our observer overly eagerly
Twitter is served as a SPA, so it is not safe to remove our root
observer after finding the first tweet list element. When the user
browses to a new page the tweet list element on page 1 may not be the
same as the element on page 2.

I've restored the old listener from pre-#96 with small changes to
accommodate the new API.